Conditioning XamlLoadPermission tests to run only under netcoreapp
