+++ src/System.Security.Permissions/tests/System.Security.Permissions.Tests.csproj
@@ -19,4 +19,7 @@
     <Compile Include="SecurityElementTests.cs" />
     <Compile Include="TrustManagerContextTests.cs" />
   </ItemGroup>
+  <ItemGroup Condition="'$(TargetGroup)' == 'netcoreapp'">
+    <Compile Include="XamlLoadPermissionTests.cs"/>
+  </ItemGroup>
 </Project>
